warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@8284 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Dropbox/src/dropbox.cpp
@@ -59,7 +59,7 @@ void CDropbox::RequestAcceessToken()
 
 		HttpRequest *request = new HttpRequest(hNetlibUser, REQUEST_POST, DROPBOX_API_URL "/oauth2/token");
 		request->pData = mir_strdup(data);
-		request->dataLength = strlen(data);
+		request->dataLength = (int)strlen(data);
 		request->AddHeader("Content-Type", "application/x-www-form-urlencoded");
 		request->AddBasicAuthHeader(DROPBOX_API_KEY, DROPBOX_API_SECRET);
